Don't base directory name on resolution parameter
This makes it possible to scale images to the desired resolution also
for multi-monitor setups.

For instance, if you have 2 1920x1200 monitors side by side, the
"dimensions directory" will be named `~/.cache/i3lock/3840x1200/`, but
you can still scale the images to 1920x1200 by supplying a
`-r 1920x1200` argument to `betterlockscreen -u`. This results in a
background image that is not stretched out over both monitors, but
instead the single image is duplicated and made to fit each monitor.

Of course, this only works if both monitors have the same resolution.